+++ vault_share/static/src/backend/vault_share.scss
